lobby game mode added
lobby game mode is in, players can join (automatically starts match for now, will change to hub area later). tweaks to character name and directory structure for cleanliness. there's also a rudimentary version of the player meshes, a body for remote players and hands for local players. the hands look like the ue4 mannequin at the moment, it's just a placeholder, I PROMISE IT'S NOT A BUG