Subclasses are working
They need padding in the ui and there are still some view target issues to sort out, but the system can load subclasses and allow the player to select between them. Stuff still needs to be cleaned up and the subclass needs to be saved to the playerstate and savegame.