Character subclasses are working
The character subclasses work now, the player spawns into the mission with the correct subclass and the groundwork is laid for adding subclass-specific abilities to the player's ASC. The subclass is also stored in the savegame, so when changing the player's class in the hub, the menu will open to the class and subclass you last used. when you select a different class, it will also default to the last subclass that was used with it.