So uh... added a lot of stuff, most of which is oriented toward the hub loadout interactable
Added an interactable interface which should cover most if not all of my interactable needs, including the groundwork for 'press' style and 'hold' style interactions. Added a hud class so that the player can tell when they are looking at an interactable object. Added an overarching parent widget class that will allow users to back out of any menu with the pause key (currently bound to tab, but can be changed later). Added the hub loadout interactable itself, which opens the class menu when the player interacts with it.