Pretty much finished projectile weapon abilities
Added reloading, mana costs, and animation handling. Took a lot of code spelunking and I now know significantly more about GAS. The architecture I've set up makes the process of adding a primary projectile ability to a character very easy, with a lot of automagic shenanigans in the background to streamline the process for any designers that work on this project (me included). I may rework how full-auto projectile abilities function, but for now it's just fine.