basic system in place
I've established a basic melee system that functions somewhere between Vermintide 2 and Mordhau. Right now it chooses the direction immediately when the player presses the button, will probably have the ability wait for an input on the mouse (if the input magnitude is less than a certain threshold) before choosing a direction.