Improved the feel of the melee input and made the UI suck less
The melee attack input feels significantly better to use now, and behaves more intuitively. When the player presses the attack button, the character will not immediately chamber an attack. Instead, it waits for a mouse movement before choosing a direction, which feels much more natural. There's also an indicator next to the crosshair that tells the player which direction their next strike will go, and it has been smoothed because it looked bad. There are still some values to tweak (namely the attack deadzone and the direction indicator smoothing), but I'm happy with the way the system functions. Now I just need to create anims for all of the subclasses.