made an in-editor tool to make melee attack profiles
further fleshing out the melee mechanics by adding customizable melee attack profiles. These define the number and angle of melee attacks, as well as the combos that are available to the player. To avoid weird behavior and bugs, I created an in-editor tool to create these profile DAs. Next will be incorporating these profiles into the melee system itself.